+++ plugins/tailwindcss/Contracts/Plugger.d.ts
@@ -1,7 +1,0 @@
-import { UIOptions } from "../types";
-import { PluginContract } from "./Plugin";
-
-export interface PluggerContract {
-    boot(options: UIOptions): void;
-    plug(plugin: PluginContract): this;
-}