+++ plugins/tailwindcss/LinkPlugin.ts
@@ -1,15 +1,11 @@
 import { Plugin } from "./Plugin";
-import { ClassName } from "./types";
+import { RequiredLinkOptions } from "./types";
 import { PluginAPI } from "tailwindcss/types/config";
 
-export class LinkPlugin extends Plugin {
-    constructor(api: PluginAPI, readonly className: ClassName) {
-        super(api);
-    }
-
-    public build() {
-        this.addComponents({
-            [this.className]: {
+export class LinkPlugin extends Plugin<RequiredLinkOptions> {
+    public create(): this {
+        return this.addComponents({
+            [this.options.linkClass]: {
                 "&:hover": {
                     textDecoration: "underline",
                 },