+++ plugins/tailwindcss/index.ts
@@ -1,12 +1,8 @@
 import Plugger from "./Plugger";
-import { UIOptions } from "./types";
-import plugin from "tailwindcss/plugin";
-import type { PluginAPI } from "tailwindcss/types/config";
 
-export const UIPlugin = plugin.withOptions((options: UIOptions = undefined) =>
-    (api: PluginAPI) => {
-        const plugger = new Plugger(api);
-        plugger.boot(options);
-    });
+export const UI = Plugger.plugUI();
+export const UIColors = Plugger.plugColors();
+export const UILinks = Plugger.plugLinks();
+export const UIEdges = Plugger.plugEdges();
 
-export default UIPlugin;
+export default UI;