+++ plugins/tailwindcss/types/index.d.ts
@@ -1,16 +1,44 @@
-import type * as Styles from "./styles";
+import type { ClassName } from "./styles";
+import type {
+    Config,
+    PluginAPI,
+    PluginCreator,
+} from "tailwindcss/types/config";
 
-export type RequiredUIOptions = {
-    linkClass: Styles.ClassName;
+export interface PluginContract<T> {
+    readonly api: PluginAPI;
+    readonly options: T;
+    create(): this;
+}
+
+export type PluginWithoutOptions =
+    | PluginCreator
+    | {
+          handler: PluginCreator;
+          config?: Partial<Config>;
+      };
+export type PluginWithOptions<T> = {
+    (options: T): {
+        handler: PluginCreator;
+        config?: Partial<Config> | undefined;
+    };
+    __isOptionsFunction: true;
 };
 
+export type RequiredLinkOptions = {
+    linkClass: ClassName;
+};
+export type LinkOptions = Partial<RequiredLinkOptions>;
+
+export type RequiredEdgeOptions = {
+    entryClass: ClassName;
+    buttonClass: ClassName;
+};
+export type EdgeOptions = Partial<RequiredEdgeOptions>;
+
+export type RequiredUIOptions = RequiredLinkOptions & RequiredEdgeOptions;
+
 export type UIOptions = Partial<RequiredUIOptions> | undefined;
-
-declare function plugin(options?: UIOptions): { handler: () => void };
-
-declare namespace plugin {
-    const __isOptionsFunction: true;
-}
 
 export type DarkModeStrategy = "class" | "media";
 export type DarkModeQuery = string;
@@ -18,5 +46,3 @@ export type DarkMode = [DarkModeStrategy, DarkModeQuery];
 
 export * from "./styles";
 export * from "./colors";
-
-export default plugin;