+++ tailwind.config.ts
@@ -1,4 +1,4 @@
-import { UIPlugin } from "./plugins/tailwindcss";
+import { UI } from "./plugins/tailwindcss";
 import { Config } from "tailwindcss";
 import defaultTheme from "tailwindcss/defaultTheme";
 
@@ -15,7 +15,7 @@ export const UIConfig: Config = {
         },
     },
 
-    plugins: [UIPlugin],
+    plugins: [UI],
 };
 
 export default UIConfig;